orders/src/events/listeners/__test__/ticket-created-listener.test.ts
import { TicketCreatedEvent } from '@rp-gittix/common';
import mongoose from 'mongoose';
import { natsWrapper } from '../../../nats-wrapper';
import { TicketCreatedListener } from '../ticket-created-listener';
import { Message } from 'node-nats-streaming';
import { Ticket } from '../../../model/ticket';

const setup = async () => {
    // Create an instance of the listener
    const listener = new TicketCreatedListener(natsWrapper.client);
    // Create a fake data event
    const data: TicketCreatedEvent['data'] = {
        id: new mongoose.Types.ObjectId().toHexString(),
        title: 'concert',
        version: 0,
        price: 20,
        userId: new mongoose.Types.ObjectId().toHexString(),
        creator: 'test',
        imageUrl: 'test',
        description: 'test',
    };
    // Create a fake message object
    // @ts-ignore
    const msg: Message = {
        ack: jest.fn().mockImplementation(() => {}),
    };
    // Return all of this stuff
    return { listener, data, msg };
};

it('creates and saves a ticket once received the ticket:created event', async () => {
    const { data, listener, msg } = await setup();
    // Call the onMessage function with the data object + message object
    await listener.onMessage(data, msg);
    // Write assertions to make sure a ticket was created!
    const ticket = await Ticket.findById(data.id);
    expect(ticket).toBeDefined();
    expect(ticket!.title).toEqual(data.title);
    expect(ticket!.price).toEqual(data.price);
});

it('acks the message received from the nats streaming server', async () => {
    const { data, listener, msg } = await setup();
    // Call the onMessage function with the data object + message object
    await listener.onMessage(data, msg);
    // Write assertions to make sure ack function is called
    expect(msg.ack).toHaveBeenCalledTimes(1);
});
